tidecorr option: change to be a bit mask
Internally the tidedisp() function accepts a bit mask for the options,
a bit to enable respective tide displacements. This had been exposed
as three options, off, solid, and solid+otl+spole. Change the exposed
option to also be a bit mask, to allow the user to enable any
combination. For example solid+spole could now be enabled. This also
sets it up to support future tide displacement options such as ocean
pole and atmospheric tide.

The gui apps are updated to just translate their current options to
this bit mask and it is left to a follow up to extend the gui apps to
be able to individually select each tide displacement option.