MAXSOLMSG: increase to 32768, to correct an overflow
Correct a further noted overflow in this buffer increasing it from
24576 to 32768, and clean up the usage of the buffer size.

A further overflow was noted, but with over 40 satellites and with
support for 5+ frequencies enabled.

This definition is used for a buffer size for solution status messages
and as the number of satellites and frequencies has increased this
buffer has needed to grow. There is a $SAT solution status line for
each satellite frequency.

In places this buffer size had been increased as overflows were noted,
but not consistently, so update the definition and correct the
references.

MAXSOLLEN - add a new definition for the solution status line buffer
length. There were cases in which the MAXSOLMSG buffer size was being
used for a single line buffer which was excessive, particularly with
the now increased size

The MAXSOLMSG definition of 32768 bytes would support 256 solution
status lines of length 128 bytes. There is an existing definition of
MAXSOLBUF as 256 which is roughly the number of solution status lines
buffered.