stream: rework low level function to return the error number
Have recv_nb() send_nb() connect_nb() accept_nb() return the error
number rather than having the caller access the error number by
calling errsock().

There were cases in which the caller was relying on the error number
when it had not been set by these functions. For example when
recv_nb() returns -1 with and error number of zero for a disconnect.

Having these function return the error number allows the caller to
rely on it and encapsulates the system differences into these
functions.